requirements.txt
django>=1.7
django-taggit>=0.12.2
django_zoneadmin>=0.1.3
django-filer>=0.9.9
django-versatileimagefield
beautifulsoup4
pygments
markdown
